refactor: 카드 등록 시의 서버 에러 객체와 타입을 api 레이어로 옮기기
